+++ packages/ti/drv/ipc/soc/V4/ipc_soc.c
@@ -499,7 +499,7 @@ int32_t Ipc_setCoreEventId(uint32_t selfId, Ipc_MbConfig* cfg, uint32_t intrCnt)
         {
             offset = range;
         }
-        outIntrBaseNum = (start + range) - offset;
+        outIntrBaseNum = (uint32_t)((start + range) - offset);
         outIntrNum = outIntrBaseNum + intrCnt;
 
         /* Translate to CorePack IRQ number */
@@ -821,7 +821,7 @@ int32_t Ipc_getIntNumRange(uint32_t coreIndex,
                 &req,
                 &res,
                 IPC_SCICLIENT_TIMEOUT);
-    if (CSL_PASS != retVal || res.range_num == 0) {
+    if ((CSL_PASS != retVal) || (res.range_num == 0U)) {
         /* Try with HOST_ID_ALL */
         req.type           = req_type[coreIndex];
         req.subtype        = (uint8_t)req_subtype[coreIndex];
